extract ghApi into shared helper [skip-e2e]
